Remove old Django tests.
