+++ .travis.yml
@@ -1,16 +1,16 @@
 sudo: false
 language: python
 env:
-  - DJANGO=2.0 TEST_DIR=jalali_test_2_0
-  - DJANGO=2.0.2 TEST_DIR=jalali_test_2_0
-  - DJANGO=2.2.3 TEST_DIR=jalali_test_2_0
+  - DJANGO=2.0 TEST_DIR=jalali_test
+  - DJANGO=2.0.2 TEST_DIR=jalali_test
+  - DJANGO=2.2.3 TEST_DIR=jalali_test
 python:
   - 3.4
   - 3.6
 jobs:
   exclude:
   - python: 3.4
-    env: DJANGO=2.2.3 TEST_DIR=jalali_test_2_0
+    env: DJANGO=2.2.3 TEST_DIR=jalali_test
 install:
   - pip install -q Django==$DJANGO
   - pip install jdatetime pytz
