+++ README.rst
@@ -187,4 +187,4 @@ From *django_jalali* version 3 and *Django* 2 you can use ``TIME_ZONE`` and ``US
 .. _settings.py: https://github.com/slashmili/django-jalali/blob/master/jalali_test/jalali_test/settings.py#L40
 .. _models.py: https://github.com/slashmili/django-jalali/blob/master/jalali_test/foo/models.py
 .. _admin.py: https://github.com/slashmili/django-jalali/blob/master/jalali_test/foo/admin.py
-.. _settings: https://github.com/slashmili/django-jalali/blob/master/jalali_test_2_0/jalali_test_2_0/settings.py#L110
+.. _settings: https://github.com/slashmili/django-jalali/blob/master/jalali_test/jalali_test/settings.py#L110
