+++ jalali_test/jalali_test/wsgi.py
@@ -4,7 +4,7 @@ WSGI config for jalali_test project.
 It exposes the WSGI callable as a module-level variable named ``application``.
 
 For more information on this file, see
-https://docs.djangoproject.com/en/1.8/howto/deployment/wsgi/
+https://docs.djangoproject.com/en/2.0/howto/deployment/wsgi/
 """
 
 import os
